rename cli folder & update readme
