+++ exe/cli/cmd/deal.go
