+++ exe/cli/cmd/deals.go
