+++ exe/cli/cmd/deals_retrieve.go
