+++ exe/cli/cmd/deals_store.go
