+++ exe/cli/cmd/deals_watch.go
