+++ exe/cli/cmd/ffs_add.go
