+++ exe/cli/cmd/ffs_info.go
