+++ exe/cli/cmd/ffs_show.go
