+++ exe/cli/cmd/slashing_get.go
