+++ exe/cli/cmd/wallet_balance.go
