+++ exe/cli/cmd/wallet_switch.go
